# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator",
        "description": "API endpoints to manage articles",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/user/password/email": {
            "post": {
                "tags": [
                    "Password Reset"
                ],
                "summary": "Send password reset link",
                "description": "Send a password reset link to the user's email",
                "operationId": "95ee4b7be754df11f150b66d9437083c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset link sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "We have e-mailed your password reset link!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email or request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid email or request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/password/reset": {
            "post": {
                "tags": [
                    "Password Reset"
                ],
                "summary": "Reset the password",
                "description": "Reset password with the provided token and new password",
                "operationId": "3dc9a242ad271344de30418f802c1996",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newPassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newPassword123"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "reset-token-here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your password has been reset successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token or password reset request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid token or password reset request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/preferences": {
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Store or update user preferences",
                "description": "This endpoint stores or updates the user preferences like providers, categories, and authors.",
                "operationId": "storeUserPreferences",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "providers",
                                    "categories",
                                    "authors"
                                ],
                                "properties": {
                                    "providers": {
                                        "description": "List of providers",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "categories": {
                                        "description": "List of categories",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "authors": {
                                        "description": "List of authors",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User preferences successfully stored or updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/personalized-feed": {
            "get": {
                "tags": [
                    "User",
                    "Articles"
                ],
                "summary": "Get personalized feed for user based on preferences",
                "description": "Fetches a personalized list of articles based on the user's preferences for sources, categories, and authors.",
                "operationId": "getUserPersonalizedFeed",
                "responses": {
                    "200": {
                        "description": "List of personalized articles returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArticleResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching articles."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserPreferenceResource": {
                "properties": {
                    "uuid": {
                        "description": "The unique identifier for the preference",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "providers": {
                        "description": "List of providers",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "authors": {
                        "description": "List of authors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categories": {
                        "description": "List of categories",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "uuid": {
                        "description": "The unique identifier for the user",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "The email of the user",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "token": {
                        "description": "The user's authentication token",
                        "type": "string",
                        "example": "your-jwt-token",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "The user account creation date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "updated_at": {
                        "description": "The last update date of the user account",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "preference": {
                        "$ref": "#/components/schemas/UserPreferenceResource"
                    }
                },
                "type": "object"
            },
            "ArticleResource": {
                "required": [
                    "uuid",
                    "title",
                    "description",
                    "source",
                    "published_at",
                    "url",
                    "category",
                    "image_url",
                    "provider",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "title": {
                        "type": "string",
                        "example": "Article Title"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a description of the article."
                    },
                    "source": {
                        "type": "string",
                        "example": "Provider Name"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/article"
                    },
                    "category": {
                        "type": "string",
                        "example": "Category Name"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "provider": {
                        "type": "string",
                        "example": "Provider Name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token to access the API endpoints",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}